+++ src/Libraries/Core/Output/ContentType.php
@@ -1,0 +1,19 @@
+<?php
+namespace Core\Output {
+	
+	/**
+	 * Interface Acceptable
+	 */
+	interface ContentType {
+		/**
+		 * @param \Core\Template $iTemplate
+		 */
+		public function __construct(private \Core\Template $iTemplate);
+
+		/**
+		 * @param string $view
+		 * @param array $data
+		 */
+		public function render(\Core\Response $iResponse);
+	}
+}