+++ src/Libraries/Core/Router.php
@@ -47,7 +47,7 @@ namespace Core {
 		 * @throws \Core\Exception\Governance
 		 * @return array
 		 */
-		public function getDefaultRoute() : array {
+		public function getConfiguredDefaultRoute() : array {
 			$defaultRoute = \Singleton::getConfiguration()->get("defaultRoute");
 
 			if(is_array($defaultRoute) !== true) {
@@ -58,13 +58,38 @@ namespace Core {
 		}
 
 		/**
+		 * Get ClassName for requests that cannot be routed
+		 * @return \Core\ClassName
+		 */
+		private function getNotFoundClassName() : ClassName {
+			return new ClassName("NotFound");
+		}
+
+		/**
+		 * Get default MethodName to be called on controlelrs
+		 * @return \Core\MethodName
+		 */
+		private function getDefaultMethodName() : MethodName {
+			return new MethodName(MethodName::DEFAULT);
+		}
+
+		/**
+		 * Handle requests that cannot be routed.
+		 * @return array An array with two elements: [0] ClassName An instance of ClassName. [1] MethodName An instance of MethodName.
+		 */
+		private function handleUnroutableRequest() : array {
+			\Core\Event::trigger("core.router.notfound", $this->request);
+			return [$this->getNotFoundClassName() , $this->getDefaultMethodName()];
+		}
+
+		/**
 		 * Returns the resolved route from path arguments
 		 * First index will always be the controller to invoke
 		 * second index, if present, will be the method name
 		 * @return array
 		 */
 		public function getRoute() : array {
-			$defaults = $this->getDefaultRoute();
+			$defaults = $this->getConfiguredDefaultRoute();
 
 			$controllerBase = $this->request->getArg(0, $defaults);
 			$methodNameBase = $this->request->getArg(1, $defaults);
@@ -76,23 +101,30 @@ namespace Core {
 				$iClassName = new ClassName($controllerBase);
 				$iReflectionClass = new \ReflectionClass($iClassName->toString());
 			} catch(\Core\Exception\FileNotFound) {
-				return [new ClassName("NotFound"), new MethodName(MethodName::DEFAULT)];
+				return $this->handleUnroutableRequest();
+			}
+
+			// Check if method name exists on class
+			// If method is not present, fallback to default
+			try {
+				$iMethodName = new MethodName($methodNameBase ?? MethodName::DEFAULT);
+				$iReflectionMethod = $iReflectionClass->getMethod($iMethodName->toString());
+			} catch(\ReflectionException) {
+				$iMethodName = $this->getDefaultMethodName();
 			}
 
 			// \Core\Exception\Governance is deliberately thrown for non-public methods
 			// as end-users could accidently end up unwanted methods if simply re-routed to MethodName::DEFAULT
 			// However we still want our fallback to MethodName::DEFAULT if the method is simply not defined.
 			try {
-				$iMethodName = new MethodName($methodNameBase ?? MethodName::DEFAULT);
 				$iReflectionMethod = $iReflectionClass->getMethod($iMethodName->toString());
 				if($iReflectionMethod->isPublic() !== true) throw new \Core\Exception\Governance;
-			} catch(\ReflectionException) {
-				$iMethodName = new \Core\MethodName(MethodName::DEFAULT);
 			} catch(\Core\Exception\Governance) {
-				$iClassName = new ClassName("NotFound");
-				$iMethodName = new MethodName(MethodName::DEFAULT);
+				return $this->handleUnroutableRequest();
 			}
 
+			\Core\Event::trigger("core.router.found", $this->request, $iClassName, $iMethodName);
+			
 			return [$iClassName, $iMethodName];
 		}
 	}